Refactor: 관심사 분리 및 주석 제거
